fix: use worker threads to batch encoding
