+++ packages/sync/messageEncoder.worker.mjs
@@ -1,0 +1,3 @@
+import { encode } from "cbor-x";
+
+export default encode;
