+++ packages/sync/package.json
@@ -23,6 +23,7 @@
   "dependencies": {
     "@mp/state": "workspace:*",
     "@mp/std": "workspace:*",
+    "piscina": "5.0.0-alpha.0",
     "cbor-x": "1.6.0",
     "ws": "8.18.0"
   },
