+++ packages/tsconfig/node.json
@@ -4,6 +4,7 @@
   "compilerOptions": {
     "skipLibCheck": true,
     "lib": ["ESNext"],
-    "types": ["node", "vite/client"]
+    "types": ["node", "vite/client"],
+    "module": "ESNext"
   }
 }
